+++ core/views.py
@@ -1,3 +1,10 @@
 from django.shortcuts import render
 
-# Create your views here.
+
+def index(request):
+
+    return render(
+        request=request,
+        template_name='index.html',
+        context={'greetings': 'Django LMS Project'}
+    )
