+++ students/templates/students/list.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'base+style.html' %}
 
 {% block title %}
     Students list
@@ -9,11 +9,32 @@
         {{form.as_p}}
         <input type="submit" value="Search">
     </form>
+<p><strong>Count of students:{{students.count}}</strong></p>
+    <table>
+        <tr>
+            <th>№</th>
+            <th>First name</th>
+            <th>Last name</th>
+            <th>Age</th>
+            <th>Birthday</th>
+            <th>Phone</th>
+            <th>📝</th>
+            <th>❌</th>
+        </tr>
 
-<br>
 
 {%for student in students%}
-    <a href="{% url 'students:delete' student.id %}">Del</a>&nbsp;
-    <a href="{% url 'students:update' student.id %}">Edit</a>&nbsp;&nbsp;{{ student }}<br>
+    <tr>
+        <td>{{forloop.counter}}</td>
+        <td>{{student.first_name}}</td>
+        <td>{{student.last_name}}</td>
+        <td>{{student.age}}</td>
+        <td>{{student.birthday}}</td>
+        <td>{{student.phone_number}}</td>
+        <td><a href="{% url 'students:update' student.id %}">Edit</a></td>
+        <td><a href="{% url 'students:delete' student.id %}">Del</a></td>
+    </tr>
+
 {% endfor %}
 {% endblock %}
+    </table>
