+++ templates/base+style.html
@@ -1,0 +1,21 @@
+{% extends 'base.html' %}
+
+{% block style_options %}
+
+  table {
+    font-family: arial, sans-serif;
+    border-collapse: collapse;
+    width: 80%;
+  }
+
+  td, th {
+    border: 1px solid #dddddd;
+    text-align: left;
+    padding: 8px;
+  }
+
+  tr:nth-child(even) {
+    background-color: #dddddd;
+  }
+
+{% endblock %}
