+++ templates/base.html
@@ -3,6 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>LMS - {% block title %}{% endblock %}</title>
+    <style>{% block style_options %}{% endblock %}</style>
+
 </head>
 <body>
     <table>
